Removed modal windows and added new modal and button components
